+++ cliutil/earthlydir.go
@@ -8,7 +8,7 @@ import (
 	"strconv"
 	"sync"
 
-	"github.com/earthly/earthly/fileutil"
+	"github.com/earthly/earthly/util/fileutil"
 	"github.com/pkg/errors"
 )
